fixed WSGI startup so that database migrations run on app initialization
