+++ mslib/mscolab/mscolab.py
@@ -39,10 +39,9 @@ from pathlib import Path
 
 from mslib import __version__
 from mslib.mscolab import migrations
-from mslib.mscolab.app import APP
+from mslib.mscolab.app import APP, create_files
 from mslib.mscolab.seed import seed_data, add_user, add_all_users_default_operation, \
     add_all_users_to_all_operations, delete_user
-from mslib.mscolab.utils import create_files
 from mslib.utils import setup_logging
