+++ mslib/mscolab/server.py
@@ -24,113 +24,18 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 """
-import sys
 import logging
-import datetime
 import socketio
 import sqlalchemy.exc
-import flask_migrate
 import hashlib
 
 from flask import jsonify, request, current_app
-from flask_mail import Mail
 from flask_cors import CORS
 from flask_httpauth import HTTPBasicAuth
 
 from mslib.mscolab.app import create_app, APP
 from mslib.mscolab.models import User
 from mslib.mscolab.sockets_manager import _setup_managers
-from mslib.mscolab.utils import create_files
-from mslib.mscolab import migrations
-
-
-def _handle_db_upgrade():
-    from mslib.mscolab.models import db
-
-    create_files()
-    inspector = sqlalchemy.inspect(db.engine)
-    existing_tables = inspector.get_table_names()
-    if ("alembic_version" not in existing_tables and len(existing_tables) > 0) or (
-        "alembic_version" in existing_tables
-        and len(existing_tables) > 1
-        and db.session.execute(sqlalchemy.text("SELECT * FROM alembic_version")).first() is None
-    ):
-        sys.exit(
-            """Your database contains no alembic_version revision identifier, but it has a schema. This suggests \
-that you have a pre-existing database but haven't followed the database migration instructions. To prevent damage to \
-your database MSColab will abort. Please follow the documentation for a manual database migration from MSColab v8/v9."""
-        )
-
-    is_empty_database = len(existing_tables) == 0 or (
-        len(existing_tables) == 1
-        and "alembic_version" in existing_tables
-        and db.session.execute(sqlalchemy.text("SELECT * FROM alembic_version")).first() is None
-    )
-    # If a database connection to migrate from is set and the target database is empty, then migrate the existing data
-    if is_empty_database and APP.config['SQLALCHEMY_DATABASE_URI_TO_MIGRATE_FROM'] is not None:
-        logging.info("The target database is empty and a database to migrate from is set, starting the data migration")
-        source_engine = sqlalchemy.create_engine(APP.config['SQLALCHEMY_DATABASE_URI_TO_MIGRATE_FROM'])
-        source_metadata = sqlalchemy.MetaData()
-        source_metadata.reflect(bind=source_engine)
-        # Determine the previous MSColab version based on the database content and upgrade to the corresponding revision
-        if "authentication_backend" in source_metadata.tables["users"].columns:
-            # It should be v9
-            flask_migrate.upgrade(directory=migrations.__path__[0], revision="c171019fe3ee")
-        else:
-            # It's probably v8
-            flask_migrate.upgrade(directory=migrations.__path__[0], revision="92eaba86a92e")
-        # Copy over the existing data
-        target_engine = sqlalchemy.create_engine(APP.config['SQLALCHEMY_DATABASE_URI'])
-        target_metadata = sqlalchemy.MetaData()
-        target_metadata.reflect(bind=target_engine)
-        with source_engine.connect() as src_connection, target_engine.connect() as target_connection:
-            for table in source_metadata.sorted_tables:
-                if table.name == "alembic_version":
-                    # Do not migrate the alembic_version table!
-                    continue
-                logging.debug("Copying table %s", table.name)
-                stmt = target_metadata.tables[table.name].insert()
-                for row in src_connection.execute(table.select()):
-                    logging.debug("Copying row %s", row)
-                    row = tuple(
-                        r.replace(tzinfo=datetime.timezone.utc) if isinstance(r, datetime.datetime) else r for r in row
-                    )
-                    target_connection.execute(stmt.values(row))
-            target_connection.commit()
-            if target_engine.name == "postgresql":
-                # Fix the databases auto-increment sequences, if it is a PostgreSQL database
-                # For reference, see: https://wiki.postgresql.org/wiki/Fixing_Sequences
-                logging.info("Using a PostgreSQL database, will fix up sequences")
-                cur = target_connection.execute(sqlalchemy.text(r"""
-SELECT
-    'SELECT SETVAL(' ||
-    quote_literal(quote_ident(sequence_namespace.nspname) || '.' || quote_ident(class_sequence.relname)) ||
-    ', COALESCE(MAX(' ||quote_ident(pg_attribute.attname)|| '), 1) ) FROM ' ||
-    quote_ident(table_namespace.nspname)|| '.'||quote_ident(class_table.relname)|| ';'
-FROM pg_depend
-    INNER JOIN pg_class AS class_sequence
-        ON class_sequence.oid = pg_depend.objid
-            AND class_sequence.relkind = 'S'
-    INNER JOIN pg_class AS class_table
-        ON class_table.oid = pg_depend.refobjid
-    INNER JOIN pg_attribute
-        ON pg_attribute.attrelid = class_table.oid
-            AND pg_depend.refobjsubid = pg_attribute.attnum
-    INNER JOIN pg_namespace as table_namespace
-        ON table_namespace.oid = class_table.relnamespace
-    INNER JOIN pg_namespace AS sequence_namespace
-        ON sequence_namespace.oid = class_sequence.relnamespace
-ORDER BY sequence_namespace.nspname, class_sequence.relname;
-"""))
-                for stmt, in cur.all():
-                    target_connection.execute(sqlalchemy.text(stmt))
-                target_connection.commit()
-        logging.info("Data migration finished")
-
-    # Upgrade to the latest database revision
-    flask_migrate.upgrade(directory=migrations.__path__[0])
-
-    logging.info("Database initialised successfully!")
 
 
 APP = create_app(imprint=APP.config['IMPRINT'], gdpr=APP.config['GDPR'])
@@ -215,9 +120,6 @@ def error413(error):
 
 
 def start_server(app, sockio, cm, fm, port=8083):
-    with app.app_context():
-        _handle_db_upgrade()
-    Mail(app)
     sockio.run(app, port=port, debug=APP.config['DEBUG'])
