+++ mslib/mscolab/utils.py
@@ -23,8 +23,6 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 """
-from pathlib import Path
-from mslib.mscolab.app import APP
 
 
 def get_recent_op_id(fm, user):
@@ -62,9 +60,3 @@ def get_message_dict(message):
         "replies": [],
         "time": message.created_at.isoformat()
     }
-
-
-def create_files():
-    Path(APP.config['OPERATIONS_DATA']).mkdir(parents=True, exist_ok=True)
-    Path(APP.config['UPLOAD_FOLDER']).mkdir(parents=True, exist_ok=True)
-    Path(APP.config['SSO_DIR']).mkdir(parents=True, exist_ok=True)
